Add rekor-mysql ExternalSecret
Add a kubectl_manifest terraform resource to create an ExternalSecret
for rekor to use to connect to the searchindexes MySQL database when
using MySQL as the index storage backend.

The username and password are the same as for the trillian database, but
they need to be accessible from the rekor-system namespace, so we create
another ExternalSecret but it can pull from the same GCP secret.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>