Add documentation on generating trusted_root.json (#1174)
The TUF mirror in this scaffolding stack does not serve a
trusted_root.json, which is needed by sigstore-go to verify artifacts
signed using a private sigstore deployment. Add documentation so that
users can create this themselves using available community tooling. This
may eventually be turned into part of the GitHub action so that clients
can use the action for integration testing.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>