Remove Cloud SQL memory utilization > 90% alert
We've been getting a lot of low urgency pages for this alert (~50 in the past day).
As discussed in the oncall meeting, this isn't a particularly useful  metric to alert on since MySQL will always use a large amount of memory.
CPU usage has been normal, and we have alerts for it, and it's probably a better metric to alert on

Signed-off-by: Priya Wadhwa <priyawadhwa@priyas-mbp.lan>