macOS support for hermetic environment
The primary change is handling the network connection between Fulcio and
the fakeoidc service. On macOS, containers run in VMs and so they do not
share a network by default. On Linux, containers can connect to one
another through shared IP addresses without mucking around in
/etc/hosts. If we made a request to the fakeoidc service without
hardcoding the issuer, then the issuer would be 'localhost', and Fulcio
couldn't connect to it since it's running in a VM. The workaround is to
simply hardcode the container name as the issuer, and as before, join
the fakeoidc container to Fulcio's network.

Also updated Cosign's container and did a little comment cleanup.

Signed-off-by: Hayden B <8418760+haydentherapper@users.noreply.github.com>