Add namespace to tesseract images (#1792)
`ko build` names the image after the name of the package that it's
building, which in the case of the two tesseract binaries is just "gcp"
and "posix", with no context on what those mean. This change extends the
KO_DOCKER_REPO variable to include "tesseract" as part of the repo name,
so the full image name will have "tesseract" in it.

When browsing the scaffolding images, both images will still just appear
as "gcp" and "posix". The alternative solution to make them more
distinguishable in the web UI would be to use `--preserve-import-paths`
which would produce an image named
"github.com/transparency-dev/tesseract/cmd/tesseract/gcp", with full
name
"ghcr.io/sigstore/scaffolding/tesseract/github.com/transparency-dev/tesseract/cmd/tesseract/gcp".

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>