Replace ct_server with TesseraCT in setup (#1793)
In the setup scaffolding workflow, update Fulcio to use the Static CT
log TesseraCT instead of the Trillian-based ct_server.

This change avoids retrieving the Fulcio roots from the rootCert
endpoint, instead favoring using the known externally generated root,
which closer simulates the mechanism for distributing key material out
of band via TUF. This means the createctconfig job is not needed for
this CT log. The createtree job is also no longer needed because there
is no Trillian tree to manage.

Signed-off-by: Colleen Murphy <colleenmurphy@google.com>